Apply review recommendations
