+++ test/proxy/ERC1967/ERC1967Proxy.test.js
@@ -1,13 +1,22 @@
 const { ethers } = require('hardhat');
 
 const shouldBehaveLikeProxy = require('../Proxy.behaviour');
+const { loadFixture } = require('@nomicfoundation/hardhat-network-helpers');
+
+const fixture = async () => {
+  const [nonContractAddress] = await ethers.getSigners();
+
+  const implementation = await ethers.deployContract('DummyImplementation');
+
+  const createProxy = (implementation, initData, opts) =>
+    ethers.deployContract('ERC1967Proxy', [implementation, initData], opts);
+
+  return { nonContractAddress, implementation, createProxy };
+};
 
 describe('ERC1967Proxy', function () {
-  // `undefined`, `null` and other false-ish opts will not be forwarded.
-  before(function () {
-    this.createProxy = function (implementation, initData, opts) {
-      return ethers.deployContract('ERC1967Proxy', [implementation, initData], opts);
-    };
+  beforeEach(async function () {
+    Object.assign(this, await loadFixture(fixture));
   });
 
   shouldBehaveLikeProxy();
