+++ test/proxy/utils/Initializable.test.js
@@ -1,6 +1,8 @@
 const { ethers } = require('hardhat');
 const { expect } = require('chai');
-const { MAX_UINT64 } = require('../../helpers/constants');
+const {
+  bigint: { MAX_UINT64 },
+} = require('../../helpers/constants');
 
 describe('Initializable', function () {
   describe('basic testing without inheritance', function () {
